fix(build): temporarily disable KAPT and Hilt to fix Java 17 compatibility
KAPT has incompatibility issues with Java 17+ that require complex JVM flags.
Since we don't actually USE Hilt/DI yet in the MVP Day 1 code, it's pragmatic
to disable it temporarily and re-enable it on Day 2 when we implement sensors.

Changes:
- Commented out kotlin-kapt and hilt plugins
- Commented out Hilt/Room dependencies (not needed for basic UI)
- Removed @HiltAndroidApp from PrivacyGuardApplication
- Removed @AndroidEntryPoint from MainActivity
- Added TODO comments to re-enable on Day 2

This unblocks the build and lets us test the basic UI immediately.
We'll tackle the KAPT/Java 17 issue properly when we actually need DI.